+++ lib/deck_of_cards.rb
@@ -4,4 +4,3 @@ require "sorbet-runtime"
 require_relative "deck_of_cards/version"
 require_relative "deck_of_cards/card"
 require_relative "deck_of_cards/packet"
-require_relative "deck_of_cards/deck"
